backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==42.0.2
pynacl==1.5.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Rate Limiting
slowapi==0.1.9

# Payment Integrations
stripe==8.2.0
paypalrestsdk==1.13.3
coinbase-commerce==1.0.1

# Blockchain/Crypto
web3==6.15.1
bitcoin-rpc==0.6.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pillow==10.2.0
httpx==0.27.0

# Monitoring & Logging
python-json-logger==2.0.7

# WebSocket
websockets==12.0

# CORS
python-cors==1.0.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
